Move Tuple location
